% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bias.R
\name{biasApp}
\alias{biasApp}
\title{Statistics app: Demonstrates changes in coefficients as model terms are added}
\usage{
biasApp(display = c("normal", "showcase"))
}
\description{
An app investigating model bias
}
\details{
The original model is created from the original data set and the model given
by \code{expr}. The response variable is taken to be the term on the left
side of your original formula. All other terms are considered explanatory,
and they are displayed as checkboxes in the manipulate box. By checking
these, each term is added to the new model. The old model coefficient is
plotted as a dotted line on each number line, and the new confidence
interval and coefficient is plotted in red. The user may vary the confidence
level and may vary the number of data points resampled to change the
placement and size of the confidence intervals. There is also a original
data checkbox which disables the n points to sample slider, as original data
is used. When original data is checked as well as only the terms in the
original model, the confidence intervals should align with the original
predictions. There is a random seed slider to set the random seed. By
default its initial value is random.

This applet allows users to see that a linear model will be drastically
different with different terms in it. A confidence interval is only
applicable to its own model or hypothesis test, it is not universally valid.
}
\examples{
biasApp() #if you do not want to display the code
biasApp(display = "showcase") #if you want to display the code
}
\author{
Andrew Rich (\email{andrew.joseph.rich@gmail.com})
,Daniel Kaplan (\email{kaplan@macalester.edu})
,Mengdie Wang(\email{mwang3@macalester.edu})
and Jingjing Yang(\email{jyang1@macalester.edu})
}
\keyword{statistics}

