% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eulerIntegrate.R
\name{eulerIntegrate}
\alias{eulerIntegrate}
\title{Interactive applet to Euler Integrate}
\usage{
eulerIntegrate(display = c("normal", "showcase"))
}
\description{
Known bugs: doesn't work with linear function
}
\examples{
eulerIntegrate() #if you do not want to display the code
eulerIntegrate(display = "showcase") #if you want to display the code
}
\author{
Andrew Rich (\email{andrew.joseph.rich@gmail.com}) and Daniel
Kaplan (\email{kaplan@macalester.edu})
, MengdieWang(\email{mwang3@macalester.edu})
and Jingjing Yang(\email{jyang1@macalester.edu})
}
\keyword{applied}
\keyword{math}

