% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eulerIntegrate.R
\name{eulerIntegrateApp}
\alias{eulerIntegrateApp}
\title{Interactive applet to Euler Integrate}
\usage{
eulerIntegrateApp(display = c("normal", "showcase"))
}
\description{
Interactive applet to Euler Integrate
}
\examples{
eulerIntegrateApp() #if you do not want to display the code
eulerIntegrateApp(display = "showcase") #if you want to display the code
}
\author{
Andrew Rich (\email{andrew.joseph.rich@gmail.com})
, Daniel Kaplan (\email{kaplan@macalester.edu})
, Mengdie Wang(\email{mwang3@macalester.edu})
and Jingjing Yang(\email{jyang1@macalester.edu})
}
\keyword{applied}
\keyword{math}

