% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggmap.R
\name{ggmapApp}
\alias{ggmapApp}
\title{ggmap app: Allows users to make simple maps}
\usage{
ggmapApp(display = c("normal", "showcase"))
}
\description{
ggmap app: Allows users to make simple maps
}
\examples{
ggmapApp()
ggmapApp(display = "showcase") #if you want to display code
}
\author{
:Daniel Kaplan (\email{kaplan@macalester.edu})
,Mengdie Wang(\email{mwang3@macalester.edu})
and Jingjing Yang(\email{jyang1@macalester.edu})
}
\keyword{statistics}

