% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggplotApp}
\alias{ggplotApp}
\title{ggplot app: Allows users to explore ggplot}
\usage{
ggplotApp(display = c("normal", "showcase"))
}
\description{
ggplot app: Allows users to explore ggplot
}
\examples{
ggplotApp()
ggplotApp(display = "showcase") #if you want to display code
}
\author{
:Daniel Kaplan (\email{kaplan@macalester.edu})
,Mengdie Wang(\email{mwang3@macalester.edu})
and Jingjing Yang(\email{jyang1@macalester.edu})
}
\keyword{statistics}

