% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glassPattern.R
\name{glassPattern}
\alias{glassPattern}
\title{Interactive applet for demonstrating Glass Pattern}
\usage{
glassPattern(display = c("normal", "showcase"))
}
\description{
Glass patterns are formed from superimposed random dot patterns:
an original image with a second image which has been generated through a linear or
nonlinear transformation of the original.A variety of different spatial patterns
such as circles, spirals, hyperbolae, can be perceived in the superimposed image set,
depending on the nature of the transformation between the two sets of dots. (Cited from Wikipedia)
}
\examples{
glassPattern() #if you do not want to display the code
glassPattern(display = "showcase") #if you want to display the code
}
\author{
Andrew Rich (\email{andrew.joseph.rich@gmail.com}) and Daniel
Kaplan (\email{kaplan@macalester.edu})
, MengdieWang(\email{mwang3@macalester.edu})
and Jingjing Yang(\email{jyang1@macalester.edu})
}
\keyword{calculus}

