% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gradSearch.R
\name{gradSearch}
\alias{gradSearch}
\title{Interactive applet for finding local extrema}
\usage{
gradSearch(display = c("normal", "showcase"))
}
\description{
Interactive applet for finding local extrema
}
\examples{
gradSearch() #if you do not want to display the code
gradSearch(display = "showcase") #if you want to display the code
}
\author{
Andrew Rich (\email{andrew.joseph.rich@gmail.com}) and Daniel
Kaplan (\email{kaplan@macalester.edu})
, MengdieWang(\email{mwang3@macalester.edu})
and Jingjing Yang(\email{jyang1@macalester.edu})
}
\keyword{calculus}

