% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gradSearch.R
\name{gradSearchApp}
\alias{gradSearchApp}
\title{Interactive applet for finding local extrema}
\usage{
gradSearchApp(display = c("normal", "showcase"))
}
\description{
Interactive applet for finding local extrema
}
\examples{
gradSearchApp() #if you do not want to display the code
gradSearchApp(display = "showcase") #if you want to display the code
}
\author{
Andrew Rich (\email{andrew.joseph.rich@gmail.com}),
Daniel Kaplan (\email{kaplan@macalester.edu})
, Mengdie Wang(\email{mwang3@macalester.edu})
and Jingjing Yang(\email{jyang1@macalester.edu})
}
\keyword{calculus}

