% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hist.R
\name{histApp}
\alias{histApp}
\title{Interactive applet to explore histograms}
\usage{
histApp(display = c("normal", "showcase"))
}
\description{
An applet which lets the user manipulate the number of bins and the center
break of a histogram corresponding to data given or a random normal
distribution.
}
\details{
The user may change the number of bins using the Number of Histogram bins
slider, and they may change the center point around which the bins are
centered with the center break slider. If data is null, the user may change
the random seed with the seed slider. Otherwise the seed slider is not
displayed.
}
\examples{
histApp() #if you do not want to display the code
histApp(display = "showcase") #if you want to display the code
}
\author{
Andrew Rich (\email{andrew.joseph.rich@gmail.com})
, Daniel Kaplan (\email{kaplan@macalester.edu})
, Mengdie Wang(\email{mwang3@macalester.edu})
and Jingjing Yang(\email{jyang1@macalester.edu})
}
\keyword{calculus}

