% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hypTest.R
\name{hypTest}
\alias{hypTest}
\title{Interactive display for hypothesis testing concepts.}
\usage{
hypTest(display = c("normal", "showcase"))
}
\arguments{
\item{useF}{Show the F distribution.  Default: a t-distribution for the
sample statistic.}
}
\value{
Just the graphics as a side effect.
}
\description{
Displays a sampling distribution and calculates power.
}
\details{
Allows the user to set various parameters that govern the sampling
distribution, and the significance, then calculates the power.  Displays all
this graphically.
}
\examples{
hypTest() #if you do not want to display the code
hypTest(display = "showcase") #if you want to display the code
}
\author{
Daniel Kaplan (\email{kaplan@macalester.edu})
, MengdieWang(\email{mwang3@macalester.edu})
and Jingjing Yang(\email{jyang1@macalester.edu})
}
\keyword{statistics}

