% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hypTest.R
\name{hypTestApp}
\alias{hypTestApp}
\title{Interactive display for hypothesis testing concepts.}
\usage{
hypTestApp(display = c("normal", "showcase"))
}
\description{
Displays a sampling distribution and calculates power.
}
\details{
Allows the user to set various parameters that govern the sampling
distribution, and the significance, then calculates the power.  Displays all
this graphically.
}
\examples{
hypTestApp() #if you do not want to display the code
hypTestApp(display = "showcase") #if you want to display the code
}
\author{
Daniel Kaplan (\email{kaplan@macalester.edu})
, Mengdie Wang(\email{mwang3@macalester.edu})
and Jingjing Yang(\email{jyang1@macalester.edu})
}
\keyword{statistics}

