% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/influence.R
\name{influenceApp}
\alias{influenceApp}
\title{Interactive applet to explore influential points in linear models}
\usage{
influenceApp(display = c("normal", "showcase"))
}
\description{
Displays data and a menu of modeling functions.  Finds a linear combination
of the selected modeling functions, with the user setting nonlinear
parameters manually.
}
\details{
The linear model described by the \code{expr} term is found and fitted
values are plotted against the data. Using the Dep. variable to plot picker
the user may choose which variable to plot on the x axis. The user can add a
new point to the data set by checking the Add new point checkbox. Using the
x and y sliders below it, the user can change the position of the new point.
The sliders' units are standard deviations from the mean. The type of
influence picker is currently nonfunctional.
}
\examples{
influenceApp() #if you do not want to display the code
influenceApp(display = "showcase") #if you want to display the code
}
\author{
Andrew Rich (\email{andrew.joseph.rich@gmail.com}), Daniel Kaplan
(\email{kaplan@macalester.edu}), Randall Pruim (\email{rpruim@calvin.edu})
, Mengdie Wang(\email{mwang3@macalester.edu})
and Jingjing Yang(\email{jyang1@macalester.edu})
}
\keyword{statistics}

