% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CopyOfmHypTest.R
\name{mHypTest}
\alias{mHypTest}
\title{Interactive display for hypothesis testing concepts.}
\usage{
mHypTest(useF = FALSE)
}
\arguments{
\item{useF}{Show the F distribution.  Default: a t-distribution for the
sample statistic.}
}
\value{
Just the graphics as a side effect.
}
\description{
Displays a sampling distribution and calculates power.
}
\details{
Allows the user to set various parameters that govern the sampling
distribution, and the significance, then calculates the power.  Displays all
this graphically.
}
\examples{
if(require(manipulate)){
		mHypTest()
		mHypTest(useF=TRUE)
	}
}
\author{
Daniel Kaplan (\email{kaplan@macalester.edu})
}
\keyword{statistics}

