% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taylor.R
\name{mTaylor}
\alias{mTaylor}
\title{Interactive Applet to Examine Taylor and Least Squares Approximations}
\usage{
mTaylor(expr, xlim = c(-5, 5), ...)
}
\arguments{
\item{expr}{a mathematical expression. See mosaic \code{D}.}

\item{xlim}{The range of the dependent variable.}

\item{\dots}{Additional arguments to \code{expr}.}
}
\description{
This applet displays a plot of a function, its Taylor series approximation,
and a least squres approximation.
}

