% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phaseLine.R
\name{phaseLine}
\alias{phaseLine}
\title{An interactive app that allows to explore phaseLine}
\usage{
phaseLine(display = c("normal", "showcase"))
}
\description{
An interactive app that allows to explore phaseLine
}
\examples{
phaseLine() #if you do not want to display the code
phaseLine(display = "showcase") #if you want to display the code
}
\author{
Andrew Rich (\email{andrew.joseph.rich@gmail.com}), Daniel Kaplan
(\email{kaplan@macalester.edu}), Randall Pruim (\email{rpruim@calvin.edu})
, MengdieWang(\email{mwang3@macalester.edu})
and Jingjing Yang(\email{jyang1@macalester.edu})
}
\keyword{differential}
\keyword{equation}

