% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phaseLine.R
\name{phaseLineApp}
\alias{phaseLineApp}
\title{An interactive app that allows to explore phaseLine}
\usage{
phaseLineApp(display = c("normal", "showcase"))
}
\description{
An interactive app that allows to explore phaseLine
}
\examples{
phaseLineApp() #if you do not want to display the code
phaseLineApp(display = "showcase") #if you want to display the code
}
\author{
Andrew Rich (\email{andrew.joseph.rich@gmail.com}), Daniel Kaplan
(\email{kaplan@macalester.edu}), Randall Pruim (\email{rpruim@calvin.edu})
, Mengdie Wang(\email{mwang3@macalester.edu})
and Jingjing Yang(\email{jyang1@macalester.edu})
}
\keyword{differential}
\keyword{equation}

