% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter}
\alias{scatter}
\title{Interactive applet to explore scatter plot}
\usage{
scatter(display = c("normal", "showcase"))
}
\description{
Interactive applet to explore scatter plot
}
\examples{
scatter() #if you do not want to display the code
scatter(display = "showcase") #if you want to display the code
}
\author{
Andrew Rich (\email{andrew.joseph.rich@gmail.com}) and Daniel
Kaplan (\email{kaplan@macalester.edu}),
, MengdieWang(\email{mwang3@macalester.edu})
}
\keyword{statistics}

