% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taylor.R
\name{taylor}
\alias{taylor}
\title{Interactive Applet to Examine Taylor and Least Squares Approximations}
\usage{
taylor(display = c("normal", "showcase"))
}
\arguments{
\item{expr}{a mathematical expression. See mosaic \code{D}.}

\item{xlim}{The range of the dependent variable.}

\item{\dots}{Additional arguments to \code{expr}.}
}
\description{
This applet displays a plot of a function, its Taylor series approximation,
and a least squres approximation.
}
\examples{
taylor() #if you do not want to display the code
taylor(display = "showcase") #if you want to display the code
}
\author{
Andrew Rich (\email{andrew.joseph.rich@gmail.com})
and Daniel Kaplan (\email{kaplan@macalester.edu})
and MengdieWang(\email{mwang3@macalester.edu})
and Jingjing Yang(\email{jyang1@macalester.edu})
}
\keyword{statistics}

