% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taylor.R
\name{taylorApp}
\alias{taylorApp}
\title{Interactive Applet to Examine Taylor and Least Squares Approximations}
\usage{
taylorApp(display = c("normal", "showcase"))
}
\description{
This applet displays a plot of a function, its Taylor series approximation,
and a least squres approximation.
}
\examples{
taylorApp() #if you do not want to display the code
taylorApp(display = "showcase") #if you want to display the code
}
\author{
Andrew Rich (\email{andrew.joseph.rich@gmail.com})
, Daniel Kaplan (\email{kaplan@macalester.edu})
, Mengdie Wang(\email{mwang3@macalester.edu})
and Jingjing Yang(\email{jyang1@macalester.edu})
}
\keyword{statistics}

